pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "insight-pilot"
version = "0.3.0"
description = "Literature research automation - search arXiv/OpenAlex, deduplicate, download PDFs, analyze and generate research reports"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Lexmount" }]
keywords = ["research", "arxiv", "openalex", "literature", "papers", "automation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "requests>=2.28.0",
    "pydantic>=2.0.0",
    "PyYAML>=6.0",
    "feedparser>=6.0.0",
    "python-dotenv>=1.0.0",
    "Jinja2>=3.1.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
pdf = [
    "PyMuPDF>=1.23.0",
    "pymupdf4llm>=0.0.5",
]

[project.scripts]
insight-pilot = "insight_pilot.cli:main"

[project.urls]
Homepage = "https://github.com/PotatoDog1669/insight-pilot"
Repository = "https://github.com/PotatoDog1669/insight-pilot"

[tool.hatch.build.targets.wheel]
packages = ["src/insight_pilot"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
